Refactor: move email related code to email package
