+++ backend/src/main/java/com/tissue/api/email/domain/EmailClient.java
@@ -1,4 +1,4 @@
-package com.tissue.api.notification.domain.service.sender;
+package com.tissue.api.email.domain;
 
 public interface EmailClient {
 	void send(String to, String subject, String body);
