+++ backend/src/main/java/com/tissue/api/email/infrastructure/DummyEmailClient.java
@@ -1,8 +1,8 @@
-package com.tissue.api.notification.infrastructure.email;
+package com.tissue.api.email.infrastructure;
 
 import org.springframework.stereotype.Component;
 
-import com.tissue.api.notification.domain.service.sender.EmailClient;
+import com.tissue.api.email.domain.EmailClient;
 
 import lombok.extern.slf4j.Slf4j;
